chore: make dbtool extensible
